Update data explorer overview page
Minimizing the number of screenshots in the docs to make it easier to maintain the page and reduce how quickly the doc images become
outdated.

Used font awesome icons in instructions to help in the case where a screenshot of the icon was removed.

Remove screenshots with markups to make it easier to replace them. Added additional details in the text now that the markup is gone.