Update data explorer docs and screenshots (#199)
* Update data explorer overview page

Minimizing the number of screenshots in the docs to make it easier to maintain the page and reduce how quickly the doc images become
outdated.

Used font awesome icons in instructions to help in the case where a screenshot of the icon was removed.

Remove screenshots with markups to make it easier to replace them. Added additional details in the text now that the markup is gone.

* Capitalize the usages of "File Explorer"

* Update data grid page

* Update filter bar page

* Update summary panel page

* Update list spacing

Co-authored-by: Julia Silge <julia.silge@gmail.com>

* Update list spacing

Co-authored-by: Julia Silge <julia.silge@gmail.com>

* Fix duplicate icon bug

* Fix list formatting

Co-authored-by: Julia Silge <julia.silge@gmail.com>

* More descriptive header

* Update filter preamble

* fix capitalization

Co-authored-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>

* Update filter preamble again

---------

Co-authored-by: Julia Silge <julia.silge@gmail.com>
Co-authored-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>